Shortcode handler and Zoom support for Google Map rendering, Flag for successful Geocode retrieval
Added a shortcode handler for rendering a static Google Map.  Also
added a zoom parameter so that the zoom level of the Google map can be
specified.  Also added a flag in Geocodable to indicate whether the
last saved address resulted in a successful Geocode retrieval.  It was
necessary because if an invalid address is entered, there is no way to
know whether the Geocode retrieval was successful because the last Lat,
Lng values successfully retrieved are maintained (I can imagine that in
some cases this is desirable, but in some applications it is important
to know whether the Lat/Lng values can be trusted).